+++ kernel/common/sys/krx_file.h
@@ -46,7 +46,7 @@ ssize_t sys_read(int fd, void *ubuf, size_t nbyte);
 ssize_t sys_write(int fd, const void *ubuf, size_t nbyte);
 int sys_getdents(int fd, void *buf, size_t nbyte);
 int sys_lseek(int fd, off_t offset, int whence, off_t *out);
-int sys_ioctl(int fd, int cmd, intptr_t arg);
+int sys_ioctl(int fd, unsigned long cmd, intptr_t arg);
 int sys_fstatat(int fd, const char *upath, int flags, struct stat *sb);
 int sys_ftruncate(int fd, off_t length);
 int sys_flock(int fd, int op);
